Change code according to PR review and advice add license header, change comment pattern, change error check pattern, use fmt.sprintf to build command line, add print partition table at begining and ending, equal or smaller size now throws no error, just a warning print and return, add partutiltest for testing env setup
