--- tools/partutil/read_partition_start_test.go
@@ -1,64 +1,0 @@
-package partutil
-
-import (
-	"testing"
-)
-
-func TestReadPartitionStart(t *testing.T) {
-	diskName := ""
-	copyFile := ""
-	t.Cleanup(func() { tearDown(copyFile, diskName) })
-	copyFile, diskName = setupFakeDisk("tmp_disk_read_start", "", t)
-
-	testData := []struct {
-		testName string
-		disk     string
-		partNum  int
-		want     int
-		expErr   bool
-	}{
-		{
-			"100KPart",
-			diskName,
-			8,
-			200,
-			false,
-		}, {
-			"InvalidDisk",
-			"./disk_file/no_disk",
-			8,
-			0,
-			true,
-		}, {
-			"InvalidPartition",
-			diskName,
-			0,
-			0,
-			true,
-		}, {
-			"NonexistPartition",
-			diskName,
-			1000,
-			0,
-			true,
-		},
-	}
-
-	for _, input := range testData {
-		t.Run(input.testName, func(t *testing.T) {
-			res, err := ReadPartitionSize(input.disk, input.partNum)
-			if (err != nil) != input.expErr {
-				if input.expErr {
-					t.Fatalf("error not found in test %s", input.testName)
-				} else {
-					t.Fatalf("error in test %s", input.testName)
-				}
-			}
-			if !input.expErr {
-				if res != input.want {
-					t.Fatalf("wrong result: %s %d to %d, exp: %d", input.disk, input.partNum, res, input.want)
-				}
-			}
-		})
-	}
-}
